This is a commit for explanation without making any change.
Suppose there is only one bucket, there is no need to handle the problem lock. Since we have to manipulate on multiple buckets, operation on multiple locks is needed. Specifically, when we want to borrow a block from other bucket , 2 locks are involved. Therefore, we should consider carefully in order of locks.